# openapi/account.yaml
openapi: 3.0.3
info:
  title: Account Service
  version: "1.0"
  description: >-
    User accounts, authentication, and saved delivery addresses.
    Consumed by commerce-bff and order-service.
paths:
  /api/v1/auth/signup:
    post:
      operationId: signup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupRequest"
      responses:
        "201":
          description: Account created and tokens issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "409":
          description: EMAIL_ALREADY_REGISTERED
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
  /api/v1/auth/login:
    post:
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Tokens issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "401":
          description: INVALID_CREDENTIALS
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
  /api/v1/auth/refresh:
    post:
      operationId: refresh
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "200":
          description: Refreshed pair; old refresh token is revoked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "401":
          description: INVALID_CREDENTIALS
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
  /api/v1/addresses:
    get:
      operationId: listAddresses
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Saved addresses for the authenticated user
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AddressResponse"
    post:
      operationId: createAddress
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddressRequest"
      responses:
        "201":
          description: Created address
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AddressResponse"
  /api/v1/addresses/{addressId}:
    get:
      operationId: getAddress
      security:
        - bearerAuth: []
      parameters:
        - name: addressId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Address detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AddressResponse"
        "404":
          description: ADDRESS_NOT_FOUND
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
    put:
      operationId: updateAddress
      security:
        - bearerAuth: []
      parameters:
        - name: addressId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddressRequest"
      responses:
        "200":
          description: Updated address
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AddressResponse"
        "404":
          description: ADDRESS_NOT_FOUND
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
    delete:
      operationId: deleteAddress
      security:
        - bearerAuth: []
      parameters:
        - name: addressId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204":
          description: Deleted
        "404":
          description: ADDRESS_NOT_FOUND
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    SignupRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 8, maxLength: 128 }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string }
    RefreshRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken: { type: string }
    TokenResponse:
      type: object
      required: [accessToken, refreshToken, tokenType, expiresInSeconds]
      properties:
        accessToken: { type: string }
        refreshToken: { type: string }
        tokenType: { type: string, example: Bearer }
        expiresInSeconds: { type: integer }
    AddressRequest:
      type: object
      required: [fullName, line1, city, postalCode, country]
      properties:
        fullName: { type: string, minLength: 1, maxLength: 120 }
        line1: { type: string, maxLength: 200 }
        line2: { type: string, maxLength: 200 }
        city: { type: string, maxLength: 100 }
        state: { type: string, maxLength: 100 }
        postalCode: { type: string, minLength: 3, maxLength: 12 }
        country: { type: string, minLength: 2, maxLength: 2 }
        phoneNumber: { type: string, maxLength: 20 }
    AddressResponse:
      type: object
      required: [id, fullName, line1, city, postalCode, country, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        fullName: { type: string }
        line1: { type: string }
        line2: { type: string, nullable: true }
        city: { type: string }
        state: { type: string, nullable: true }
        postalCode: { type: string }
        country: { type: string }
        phoneNumber: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    Problem:
      type: object
      required: [type, title, status, code, correlationId]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        code: { type: string }
        detail: { type: string }
        correlationId: { type: string }